extract StudentAccessInteractorMock into shared mocks to remove duplicates
