+++ Core/CoreTests/Features/Planner/CalendarMain/PlannerViewControllerTests.swift
@@ -188,20 +188,3 @@ class PlannerViewControllerTests: CoreTestCase {
         override var isDragging: Bool { true }
     }
 }
-
-private class StudentAccessInteractorMock: StudentAccessInteractor {
-    private let isStudentAccessRestricted: CurrentValueSubject<Bool, Never>
-
-    init(isRestricted: Bool = false) {
-        isStudentAccessRestricted = CurrentValueSubject(isRestricted)
-    }
-
-    func isRestricted() -> AnyPublisher<Bool, Never> {
-        isStudentAccessRestricted.eraseToAnyPublisher()
-    }
-
-    // Optional: allow changing value in test
-    func setRestricted(_ value: Bool) {
-        isStudentAccessRestricted.send(value)
-    }
-}
